pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "mlops_exam_project"
version = "0.0.1"
description = "Exam project for the course 02476 Machine Learning Operations"
authors = [
  { name = "Group_31", email = "your@email.com" },
]
license = { file = "LICENSE" }
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "hydra-core>=1.3.2",
    "logging>=0.4.9.6",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch==2.6.0",
    "typer==0.15.1",
    "ucimlrepo>=0.0.7",
    "uvicorn==0.34.0",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "pandas==2.3.3",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.11.5",
    "mkdocstrings-python==1.12.2",
    "torch",
    "scikit-learn"
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true


[tool.coverage.run]
omit = ["tests/*"]

[tool.ruff]
# Exclude common directories like virtual environments and build artifacts
exclude = [
".git", ".venv", "build", "dist", "node_modules"
]
